+++ src/delphyne/server/fastapi.py
@@ -35,11 +35,7 @@ def make_server(launcher: TaskLauncher):
     ):
         stream_eval = ta.stream_of_fun(analysis.evaluate_demo)
         context = context.with_root(Path(workspace_root))
-        loader = dp.ObjectLoader(
-            strategy_dirs=context.strategy_dirs,
-            modules=context.modules,
-            extra_objects=std.stdlib_globals(),
-        )
+        loader = context.object_loader(extra_objects=std.stdlib_globals())
         stream = launcher(
             request,
             fb.DemoFeedback,