+++ src/delphyne/stdlib/tasks.py
@@ -19,6 +19,7 @@ from typing import Any, Concatenate, Literal, Never, Protocol
 import delphyne.analysis.feedback as fb
 import delphyne.stdlib.tasks as ta
 import delphyne.utils.typing as ty
+from delphyne.analysis import ObjectLoader
 from delphyne.utils.yaml import pretty_yaml
 
 DEFAULT_STRATEGY_DIRS = (Path("."),)
@@ -132,6 +133,13 @@ class CommandExecutionContext:
             files.
         cache_root: The directory in which to store all request cache
             subdirectories.
+        init: A sequence of initialization functions to call
+            before any object is loaded. Each element specifies a
+            qualified function name, or a pair of a qualified function
+            name and of a dictionary of arguments to pass. An
+            initialization function must be **idempotent**: calling it
+            several times should have the same effect as calling it
+            once.
         result_refresh_period: The period in seconds at which the
             current result is computed and communicated to the UI (e.g.,
             the period at which the current trace is exported when
@@ -169,6 +177,7 @@ class CommandExecutionContext:
     prompt_dirs: Sequence[Path] = DEFAULT_PROMPTS_DIRS
     data_dirs: Sequence[Path] = DEFAULT_DATA_DIRS
     cache_root: Path | None = None
+    init: Sequence[str | tuple[str, dict[str, Any]]] = ()
     result_refresh_period: float | None = None
     status_refresh_period: float | None = None
     workspace_root: Path | None = None
@@ -184,9 +193,21 @@ class CommandExecutionContext:
             prompt_dirs=[root / d for d in self.prompt_dirs],
             data_dirs=[root / d for d in self.data_dirs],
             cache_root=None if self.cache_root is None else self.cache_root,
+            init=self.init,
             result_refresh_period=self.result_refresh_period,
             status_refresh_period=self.status_refresh_period,
             workspace_root=root,
+        )
+
+    def object_loader(
+        self, reload: bool = True, extra_objects: dict[str, Any] | None = None
+    ) -> ObjectLoader:
+        return ObjectLoader(
+            strategy_dirs=self.strategy_dirs,
+            modules=self.modules,
+            extra_objects=extra_objects,
+            reload=reload,
+            initializers=self.init,
         )